requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0

python-twitter
nltk == 3.4
requests == 2.21.0
unittest2 == 1.1.0
django 
coverage == 4.5.3
mock == 2.0.0
behave == 1.2.6
selenium
jsonschema == 2.6.0
django-bootstrap4 == 0.0.8